+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This update introduces several improvements and bug fixes.
 
 ### Added
  - Add index for listing file attachments by creation date
+ - Add configuration for maximum items in file attachments listing